+++ cpp/core.h
@@ -242,8 +242,8 @@ Tuple2<BigStr*, int>* MakeDirCacheKey(BigStr* path);
 
 namespace pyutil {
 
-float infinity();
-float nan();
+double infinity();
+double nan();
 
 bool IsValidCharEscape(BigStr* c);
 BigStr* ChArrayToString(List<int>* ch_array);
